Ban dev address from sending
Since we can't set the premine to 0 (this changes everything from block 0-100...which is everything) we need to find a way to disable the premine. During the CheckTransaction method, we can pull the MIE address and compare it to the dev. If it's the dev (and the block height is over 100,000), do not allow sending from the dev to anyone (even if he modifies the source, the consensus of the blockchain nodes will prevent the spending of it through the majority).